requirements.txt
# Info Research Agent MVP
anthropic>=0.25.0
streamlit>=1.31.0
requests>=2.31.0
python-dotenv>=1.0.0
praw>=7.7.0
